Organizacion de registro
