+++ front/src/services/Servicios/bienes.jsx
@@ -1,0 +1,5 @@
+import axios from "../../api/axios";
+
+export const postBienes = (data) => {
+    //pass
+}
